move place to core and adapt json
