--- app/src/main/assets/api/place.min.json
@@ -1,181 +1,0 @@
-{
-  "cityId": 1,
-  "cityName": "Munich",
-  "updatedAt": "2025-08-14T12:00:00Z",
-  "places": [
-    { 
-      "id": 1, 
-      "cityId": 1,
-      "name": "Marienplatz",        
-      "category": "LANDMARK",  
-      "imageUrl": "marienplatz",
-      "coords": { "lat": 48.137154, "lon": 11.575382 },
-      "address": "Marienplatz, 80331 München, Germany",
-      "openingHoursText": "Always open",
-      "priceRange": "LOW",
-      "tags": ["historic", "square", "central"],
-      "otherImages": ["marienplatz_2", "marienplatz_3"],
-      "story": {
-        "mainTitle": "The Heart of Munich",
-        "subTitle": "Historic City Center",
-        "overview": "Marienplatz is the central square in the heart of Munich's Altstadt (Old Town).",
-        "highlights": [
-          { "title": "Gothic Architecture", "text": "Home to the New Town Hall with its famous Glockenspiel." },
-          { "title": "Historic Significance", "text": "Dating back to 1158, it's been the city's main square for centuries." }
-        ]
-      },
-      "facts": [
-        { "text": "The square was named after the Mariensäule (Mary's Column) erected in 1638.", "sourceName": "Munich City Archives", "sourceUrl": "https://www.muenchen.de" }
-      ],
-      "attribution": [
-        { "text": "Photo by Munich Tourism Board", "url": "https://www.muenchen.de" }
-      ],
-      "updatedAt": "2025-08-14T12:00:00Z"
-    },
-    { 
-      "id": 2, 
-      "cityId": 1,
-      "name": "Nymphenburg Palace", 
-      "category": "LANDMARK",  
-      "imageUrl": "nymphenburg_palace",
-      "coords": { "lat": 48.158902, "lon": 11.503906 },
-      "address": "Schloss Nymphenburg 1, 80638 München, Germany",
-      "openingHoursText": "9:00 AM - 6:00 PM",
-      "priceRange": "MEDIUM",
-      "tags": ["palace", "baroque", "garden"],
-      "otherImages": ["nymphenburg_2", "nymphenburg_3"],
-      "story": {
-        "mainTitle": "Baroque Masterpiece",
-        "subTitle": "Summer Residence of Bavarian Kings",
-        "overview": "Nymphenburg Palace is one of the largest royal palaces in Europe.",
-        "highlights": [
-          { "title": "Baroque Architecture", "text": "Built in the 17th century, it showcases magnificent baroque design." },
-          { "title": "Beautiful Gardens", "text": "Surrounded by extensive gardens and parklands." }
-        ]
-      },
-      "facts": [
-        { "text": "Construction began in 1664 by Elector Ferdinand Maria.", "sourceName": "Bavarian Palace Department", "sourceUrl": "https://www.schloesser.bayern.de" }
-      ],
-      "attribution": [
-        { "text": "Photo by Bavarian Palace Department", "url": "https://www.schloesser.bayern.de" }
-      ],
-      "updatedAt": "2025-08-14T12:00:00Z"
-    },
-    { 
-      "id": 3, 
-      "cityId": 1,
-      "name": "BMW Museum",         
-      "category": "MUSEUM",    
-      "imageUrl": "bmw_museum",
-      "coords": { "lat": 48.177716, "lon": 11.559800 },
-      "address": "Am Olympiapark 2, 80809 München, Germany",
-      "openingHoursText": "10:00 AM - 6:00 PM",
-      "priceRange": "MEDIUM",
-      "tags": ["automotive", "technology", "design"],
-      "otherImages": ["bmw_2", "bmw_3"],
-      "story": {
-        "mainTitle": "Automotive Excellence",
-        "subTitle": "BMW's Heritage and Innovation",
-        "overview": "The BMW Museum showcases the company's history and automotive innovations.",
-        "highlights": [
-          { "title": "Historic Collection", "text": "Features classic BMW models and concept cars." },
-          { "title": "Interactive Exhibits", "text": "Modern displays showcasing BMW's engineering achievements." }
-        ]
-      },
-      "facts": [
-        { "text": "Opened in 1973, the museum building is shaped like a car cylinder head.", "sourceName": "BMW Group", "sourceUrl": "https://www.bmwgroup.com" }
-      ],
-      "attribution": [
-        { "text": "Photo by BMW Group", "url": "https://www.bmwgroup.com" }
-      ],
-      "updatedAt": "2025-08-14T12:00:00Z"
-    },
-    { 
-      "id": 4, 
-      "cityId": 1,
-      "name": "Alter Peter Tower",  
-      "category": "VIEWPOINT", 
-      "imageUrl": "alter_peter",
-      "coords": { "lat": 48.137154, "lon": 11.575382 },
-      "address": "Rindermarkt 1, 80331 München, Germany",
-      "openingHoursText": "9:00 AM - 5:00 PM",
-      "priceRange": "LOW",
-      "tags": ["church", "tower", "viewpoint"],
-      "otherImages": ["alter_peter_2", "alter_peter_3"],
-      "story": {
-        "mainTitle": "Munich's Oldest Church",
-        "subTitle": "St. Peter's Church Tower",
-        "overview": "Alter Peter is Munich's oldest parish church, offering panoramic city views.",
-        "highlights": [
-          { "title": "Historic Church", "text": "First mentioned in 1225, it's the city's oldest church." },
-          { "title": "Panoramic Views", "text": "Climb the tower for breathtaking views of Munich." }
-        ]
-      },
-      "facts": [
-        { "text": "The church was first mentioned in documents from 1225.", "sourceName": "Munich City Archives", "sourceUrl": "https://www.muenchen.de" }
-      ],
-      "attribution": [
-        { "text": "Photo by Munich Tourism Board", "url": "https://www.muenchen.de" }
-      ],
-      "updatedAt": "2025-08-14T12:00:00Z"
-    },
-    { 
-      "id": 5, 
-      "cityId": 1,
-      "name": "Kanso Coffee",       
-      "category": "COFFEE",    
-      "imageUrl": "kanso",
-      "coords": { "lat": 48.137154, "lon": 11.575382 },
-      "address": "Kaufingerstraße 15, 80331 München, Germany",
-      "openingHoursText": "7:00 AM - 7:00 PM",
-      "priceRange": "MEDIUM",
-      "tags": ["coffee", "specialty", "artisan"],
-      "otherImages": ["kanso_2", "kanso_3"],
-      "story": {
-        "mainTitle": "Artisan Coffee Experience",
-        "subTitle": "Specialty Coffee in Munich",
-        "overview": "Kanso Coffee offers high-quality, carefully sourced coffee beans and expert brewing.",
-        "highlights": [
-          { "title": "Quality Beans", "text": "Sourced from sustainable farms around the world." },
-          { "title": "Expert Brewing", "text": "Baristas trained in various brewing methods." }
-        ]
-      },
-      "facts": [
-        { "text": "Founded in 2018, Kanso focuses on single-origin coffee beans.", "sourceName": "Kanso Coffee", "sourceUrl": "https://www.kansocoffee.de" }
-      ],
-      "attribution": [
-        { "text": "Photo by Kanso Coffee", "url": "https://www.kansocoffee.de" }
-      ],
-      "updatedAt": "2025-08-14T12:00:00Z"
-    },
-    { 
-      "id": 6, 
-      "cityId": 1,
-      "name": "Hofbräuhaus",        
-      "category": "FOOD",      
-      "imageUrl": "hofbraeuhaus",
-      "coords": { "lat": 48.137154, "lon": 11.575382 },
-      "address": "Platzl 9, 80331 München, Germany",
-      "openingHoursText": "9:00 AM - 11:30 PM",
-      "priceRange": "MEDIUM",
-      "tags": ["beer", "traditional", "bavarian"],
-      "otherImages": ["hofbraeuhaus_2", "hofbraeuhaus_3"],
-      "story": {
-        "mainTitle": "Bavarian Beer Tradition",
-        "subTitle": "Historic Beer Hall",
-        "overview": "Hofbräuhaus is one of Munich's most famous beer halls, serving traditional Bavarian cuisine.",
-        "highlights": [
-          { "title": "Historic Setting", "text": "Founded in 1589, it's one of Munich's oldest beer halls." },
-          { "title": "Traditional Cuisine", "text": "Authentic Bavarian dishes and world-famous beer." }
-        ]
-      },
-      "facts": [
-        { "text": "Founded in 1589 by Wilhelm V, Duke of Bavaria.", "sourceName": "Hofbräuhaus", "sourceUrl": "https://www.hofbraeuhaus.de" }
-      ],
-      "attribution": [
-        { "text": "Photo by Hofbräuhaus", "url": "https://www.hofbraeuhaus.de" }
-      ],
-      "updatedAt": "2025-08-14T12:00:00Z"
-    }
-  ]
-}
